nonstandard_cluster_project/requirements.txt
numpy>=1.24
pandas>=2.0
tqdm>=4.66
opencv-python>=4.8
Pillow>=10.0
scikit-learn>=1.3
pyyaml>=6.0
umap-learn>=0.5.5
# 如果用 HDBSCAN, 需要安装这个库
hdbscan>=0.8.33